+++ ballcat-admin/ballcat-admin-core/src/main/java/com/hccake/ballcat/admin/modules/lov/typehandler/ListLovSelectOptionTypeHandler.java
@@ -3,7 +3,7 @@ package com.hccake.ballcat.admin.modules.lov.typehandler;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hccake.ballcat.admin.modules.lov.model.entity.LovSelectOptions;
-import com.hccake.ballcat.common.core.util.SpringUtil;
+import com.hccake.ballcat.common.conf.util.SpringUtils;
 import org.apache.ibatis.type.JdbcType;
 import org.apache.ibatis.type.TypeHandler;
 
@@ -23,7 +23,7 @@ public class ListLovSelectOptionTypeHandler implements TypeHandler<List<LovSelec
 
 	public ObjectMapper getOm() {
 		if (om == null) {
-			om = SpringUtil.getBean(ObjectMapper.class);
+			om = SpringUtils.getBean(ObjectMapper.class);
 		}
 		return om;
 	}