+++ ballcat-common/ballcat-common-conf/src/main/java/com/hccake/ballcat/common/conf/config/JsonConfig.java
@@ -3,7 +3,8 @@ package com.hccake.ballcat.common.conf.config;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hccake.ballcat.common.core.jackson.JavaTimeModule;
 import com.hccake.ballcat.common.core.jackson.NullSerializerModifier;
-import com.hccake.ballcat.common.core.util.JacksonUtils;
+import com.hccake.ballcat.common.util.JsonUtils;
+import com.hccake.ballcat.common.util.json.JacksonAdapter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
@@ -15,24 +16,25 @@ import org.springframework.context.annotation.Configuration;
  * @date 2019/10/17 22:14
  */
 @Configuration
-@ConditionalOnClass(ObjectMapper.class)
-public class JacksonConfig {
+public class JsonConfig {
 
 	/**
 	 * 自定义objectMapper
 	 * @return ObjectMapper
 	 */
 	@Bean
+	@ConditionalOnClass(ObjectMapper.class)
 	@ConditionalOnMissingBean(ObjectMapper.class)
 	public ObjectMapper objectMapper() {
-		JacksonUtils.config(mapper -> {
+		JacksonAdapter adapter = (JacksonAdapter) JsonUtils.getAdapter();
+		adapter.config(mapper -> {
 			// NULL值修改
 			mapper.setSerializerFactory(
 					mapper.getSerializerFactory().withSerializerModifier(new NullSerializerModifier()));
 			// 时间解析器
 			mapper.registerModule(new JavaTimeModule());
 		});
-		return JacksonUtils.getMapper();
+		return JacksonAdapter.getMapper();
 	}
 
 }