+++ ballcat-common/ballcat-common-conf/src/main/java/com/hccake/ballcat/common/conf/util/SpringUtils.java
@@ -1,4 +1,4 @@
-package com.hccake.ballcat.common.core.util;
+package com.hccake.ballcat.common.conf.util;
 
 import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;
@@ -12,13 +12,13 @@ import java.util.Map;
  * @author lingting 2020/6/12 16:36
  */
 @Component
-public class SpringUtil implements ApplicationContextAware {
+public class SpringUtils implements ApplicationContextAware {
 
 	private static ApplicationContext context;
 
 	@Override
-	public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
-		context = applicationContext;
+	public void setApplicationContext(ApplicationContext context) throws BeansException {
+		SpringUtils.context = context;
 	}
 
 	@SuppressWarnings("unchecked")